pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "pydantic[email]>=2.11.4",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
